requirements.txt
celery==3.1.25
Django==1.10.5
django-celery==3.2.1
pytz==2016.10
PyYAML==3.12
requests==2.13.0
